feat(filters): remove atomic blocks instead of making them unstyled
