+++ src/lib/filters/atomic.test.js
@@ -1,0 +1,219 @@
+import { EditorState, convertFromRaw } from "draft-js"
+
+import {
+  preserveAtomicBlocks,
+  resetAtomicBlocks,
+  removeInvalidAtomicBlocks,
+} from "./atomic"
+
+describe("atomic", () => {
+  describe("#preserveAtomicBlocks", () => {
+    it("works", () => {
+      const content = convertFromRaw({
+        entityMap: {
+          "4": {
+            type: "IMAGE",
+            data: {
+              src: "/example.png",
+            },
+          },
+          "5": {
+            type: "EMBED",
+            data: {
+              url: "http://www.youtube.com/watch?v=y8Kyi0WNg40",
+            },
+          },
+        },
+        blocks: [
+          {
+            key: "a",
+            text: " ",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+          },
+          {
+            key: "b",
+            text: "📷",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+          },
+          {
+            key: "c",
+            text: " ",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 5,
+              },
+            ],
+          },
+          {
+            key: "d",
+            text: " ",
+            entityRanges: [],
+          },
+          {
+            key: "e",
+            text: "📷 Star list",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+          },
+        ],
+      })
+
+      expect(
+        preserveAtomicBlocks(["IMAGE"], content)
+          .getBlockMap()
+          .map((b) => b.getType())
+          .toJS(),
+      ).toEqual({
+        a: "atomic",
+        b: "atomic",
+        c: "unstyled",
+        d: "unstyled",
+        e: "unstyled",
+      })
+    })
+
+    it("no normalisation = no change", () => {
+      const content = EditorState.createEmpty().getCurrentContent()
+      expect(preserveAtomicBlocks([], content)).toBe(content)
+    })
+  })
+
+  describe("#resetAtomicBlocks", () => {
+    it("works", () => {
+      const content = convertFromRaw({
+        entityMap: {
+          "4": {
+            type: "IMAGE",
+            data: {
+              src: "/example.png",
+            },
+          },
+        },
+        blocks: [
+          {
+            key: "a",
+            text: "📷",
+            type: "atomic",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+          },
+          {
+            key: "b",
+            text: " ",
+            type: "atomic",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+            inlineStyleRanges: [
+              {
+                offset: 0,
+                length: 1,
+                style: "BOLD",
+              },
+            ],
+          },
+        ],
+      })
+
+      expect(
+        resetAtomicBlocks(content)
+          .getBlockMap()
+          .map((b) => ({
+            text: b.getText(),
+            type: b.getType(),
+            style: b.getInlineStyleAt(0).size,
+          }))
+          .toJS(),
+      ).toEqual({
+        a: { text: " ", type: "atomic", style: 0 },
+        b: { text: " ", type: "atomic", style: 0 },
+      })
+    })
+
+    it("no filtering = no change", () => {
+      const content = EditorState.createEmpty().getCurrentContent()
+      expect(resetAtomicBlocks(content)).toBe(content)
+    })
+  })
+
+  describe("#removeInvalidAtomicBlocks", () => {
+    it("works", () => {
+      const content = convertFromRaw({
+        entityMap: {
+          "4": {
+            type: "IMAGE",
+            data: {
+              src: "/example.png",
+            },
+          },
+          "5": {
+            type: "EMBED",
+            data: {
+              url: "http://www.youtube.com/watch?v=y8Kyi0WNg40",
+            },
+          },
+        },
+        blocks: [
+          {
+            key: "a",
+            text: " ",
+            type: "atomic",
+            entityRanges: [
+              {
+                offset: 0,
+                length: 1,
+                key: 4,
+              },
+            ],
+            inlineStyleRanges: [],
+          },
+        ],
+      })
+
+      expect(
+        removeInvalidAtomicBlocks([{ type: "IMAGE" }], content)
+          .getBlockMap()
+          .map((b) => ({
+            text: b.getText(),
+            type: b.getType(),
+          }))
+          .toJS(),
+      ).toEqual({
+        a: { text: " ", type: "atomic" },
+      })
+    })
+
+    it("no filtering = no change", () => {
+      const content = EditorState.createEmpty().getCurrentContent()
+      expect(removeInvalidAtomicBlocks([], content)).toBe(content)
+    })
+  })
+})
