+++ src/lib/filters/blocks.test.js
@@ -14,96 +14,6 @@ import {
 } from "./blocks"
 
 describe("blocks", () => {
-  describe("#preserveAtomicBlocks", () => {
-    it("works", () => {
-      const content = convertFromRaw({
-        entityMap: {
-          "4": {
-            type: "IMAGE",
-            data: {
-              src: "/example.png",
-            },
-          },
-          "5": {
-            type: "EMBED",
-            data: {
-              url: "http://www.youtube.com/watch?v=y8Kyi0WNg40",
-            },
-          },
-        },
-        blocks: [
-          {
-            key: "a",
-            text: " ",
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 4,
-              },
-            ],
-          },
-          {
-            key: "b",
-            text: "📷",
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 4,
-              },
-            ],
-          },
-          {
-            key: "c",
-            text: " ",
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 5,
-              },
-            ],
-          },
-          {
-            key: "d",
-            text: " ",
-            entityRanges: [],
-          },
-          {
-            key: "e",
-            text: "📷 Star list",
-            entityRanges: [
-              {
-                offset: 0,
-                length: 1,
-                key: 4,
-              },
-            ],
-          },
-        ],
-      })
-
-      expect(
-        preserveAtomicBlocks([IMAGE], content)
-          .getBlockMap()
-          .map((b) => b.getType())
-          .toJS(),
-      ).toEqual({
-        a: "atomic",
-        b: "atomic",
-        c: "unstyled",
-        d: "unstyled",
-        e: "unstyled",
-      })
-    })
-
-    it("no normalisation = no change", () => {
-      const content = EditorState.createEmpty().getCurrentContent()
-      expect(preserveAtomicBlocks([], content)).toBe(content)
-    })
-  })
-
   describe("#removeInvalidDepthBlocks", () => {
     it("normalises depth to a given number", () => {
       const content = convertFromRaw({
